src/CognitiveSearch.Skills/Python/Vision/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
azure-cognitiveservices-vision-computervision
azure-ai-formrecognizer>=3.1.0
Pillow
ocrlayout==0.8
